requirements.txt
psycopg == 3.2.3
psycopg2-binary == 2.9.10